Add rate limiter on scim endpoints
